web/requirements.txt
flask==2.3.3
boto3==1.28.38
pandas==2.0.3
pyarrow==13.0.0
python-dotenv==1.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.6
SQLAlchemy==2.0.15
Flask-SQLAlchemy==3.0.5
alembic==1.11.1 